Prevent instrumentation of the tracer setup by default
For example, the Jaeger tracer uses OkHttp in its HTTP sender, so if
instrumenting a project with that library and letting the TracerResolver
initialize the tracer, an error will occur since the synchronization
doesn't apply to recursive invocations of the initTracer method on the
same thread.

This can be enabled with a special system property if so desired.